+++ ckks/ckks_test.go
@@ -971,7 +971,7 @@ func testDecryptPublic(tc *testContext, t *testing.T) {
 			t.Skip("method is unsuported when params.PCount() == 0")
 		}
 
-		if tc.params.MaxLevel() < 5 {
+		if tc.params.MaxLevel() < 6 {
 			t.Skip("skipping test for params max level < 5")
 		}
 
@@ -979,7 +979,7 @@ func testDecryptPublic(tc *testContext, t *testing.T) {
 
 		values, _, ciphertext := newTestVectors(tc, tc.encryptorSk, complex(-1, 0), complex(1, 0), t)
 
-		poly := Approximate(cmplx.Sin, complex(-1.5, 0), complex(1.5, 0), 15)
+		poly := Approximate(cmplx.Sin, complex(-1.5, 0), complex(1.5, 0), 31)
 
 		for i := range values {
 			values[i] = cmplx.Sin(values[i])